// sdk/lazorkit-ts/src/generated/accounts/authorityAccount.ts
export type AuthorityAccount = {
  discriminator: number;
  authorityType: number;
  role: number;
  bump: number;
  version: number;
  padding: ReadonlyUint8Array;
  counter: bigint;
  wallet: Address;
};