// sdk/lazorkit-ts/src/generated/accounts/sessionAccount.ts
export type SessionAccount = {
  discriminator: number;
  bump: number;
  version: number;
  padding: ReadonlyUint8Array;
  wallet: Address;
  sessionKey: Address;
  expiresAt: bigint;
};